fix unset

+++ srcs/builtins/unset.c
@@ -61,7 +61,7 @@ void	secret_unset(char **a, t_mini *mini)
 		while (a[i])
 		{
 			if (ft_strncmp(a[i], env->next->value, \
-				env_name_size(env->next->value)) == 0)
+				env_name_size(env->next->value)) == 0 && (ft_strlen(a[i]) == ft_strlen(env->next->value) || env->next->value[ft_strlen(a[i])] == '='))
 			{
 				tmp = env->next->next;
 				delete_node(mini, env->next);
@@ -84,14 +84,13 @@ int	ft_unset(char **a, t_mini *mini)
 	if (!(a[1]))
 		return (SUCCESS);
 	delete_first_node(mini, env, a, 0);
-	i = 1;
 	while (env && env->next)
 	{
 		i = 1;
 		while (a[i])
 		{
 			if (ft_strncmp(a[i], env->next->value, \
-				env_name_size(env->next->value)) == 0)
+				env_name_size(env->next->value)) == 0 && ft_strlen(a[i]) == env_name_size(env->next->value))
 			{
 				tmp = env->next->next;
 				delete_node(mini, env->next);
